Version bumped wvdial.  Fixed wvstreams to use the latest available release 3.74.0.  Added a DEPENDS file just full of optional_depends (hey it's a LIBRARY and there is no getting around all choices!)
Made db a required DEPENDS (the devs have it listed as optional) as
wvstreams-3.74.0 will NOT compile at all unless Berkely DB is present
and I even had to hack their vars.mk file as their configure script is borked
and does not correctly detect that Berkely DB3 is present and working on
a Lunar box.

Compiled and tested with gcc-3.2.3, glibc-2.3.2.


git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@4646 fe636330-a7f5-0310-8408-fdd8893149bb