Okay, here is the reworked mozilla module.  SVG is also disable here, just like firefox and thunderbird.  Look for svg to be re-enabled whenever moz-devs release firefox-1.1.
I also fixed the extensions so that gnome-vfs is NOT required.  I have tested
this on a lunar box with no gnome-vfs2 installed and it does compile without
gnome-vfs.  That should make the gnome hater much happier now that mozilla
does NOT pull in most all of gnome2 anymore.

Version bumped to mozilla-1.7.11 (finally).


git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@16261 fe636330-a7f5-0310-8408-fdd8893149bb