Reverting this module back to check-0.6.2.  Too little testing by the lunar dev that bumped this module to 0.6.3.  check-0.6.3. fails to compile on a standard Lunar box (i.e. 1.5.1 ISO, with gcc-3.4.4, glibc-2.3.5, kernel 2.4.31).
check-0.6.3, also BREAKS discover-2.0.7 -- which is the MAIN reason why the
check module is even in moonbase.  check-0.6.3, can be made to compile by
removing the "make check" step from the BUILD, BUT the resulting module still
breaks the discover module so check-0.6.3 is basically USELESS for Lunar in
its current state.


git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@16441 fe636330-a7f5-0310-8408-fdd8893149bb