There were a few more changes then anticipated.
Need the gnokiirc from docs unless you want to build one from
scratch. Makes a group gnokii if your so inclinded to use one
and sets locking to no, else the thing won't start complaining 
about a lock file.

The gnokiirc *may be* needed in a users ~/.gnokiirc dir. 

At this point it lanunches but I do not have a device to test fully
with, so if additional tweaking is needed just holler.



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@17554 fe636330-a7f5-0310-8408-fdd8893149bb