Converted alsa driver to use asoundlib instead of directly talking to kernel driver.  This also means that device names are now the ALSA logical names instead of /dev type names.
Added ALSA support to play/rec scripts.

When seek()ing in AIFF, SMP, and WAV handlers, remaining samples were
incorrectly computed based on bytes and not samples.

And more. Based on the ALSA stuff above I have removed the --diable-alsa-dsp
if I proves to be a problem I will add it.

Compile with gcc-3.4.4 and glibc-2.3.6



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@17577 fe636330-a7f5-0310-8408-fdd8893149bb