Initial import from submissions.
A couple of notes. 

1. POST_INSTALL does what they refer to as a soft install which should not be a
problem for now during the current series of postgresql. When a major revision 
of postgresql is done this soft install may fail, that is the createlang may
not alter the appropriate tables. Right now this is not of any concern. I will
cross that bridge when it arrives.

2. If you have passworded the user "postgis" the install will halt waiting
for you to supply that password. I do not know of a way around this.

Compiled with gcc-3.4.5, glibc-2.3.6 and 2.6.15-rc7



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@17901 fe636330-a7f5-0310-8408-fdd8893149bb