There is a typo that gets inserted by "unix/tcl.m4" into "configure" resulting in this error when linned;
checking system version (for dynamic loading)...
../configure: line 7653: syntax error near unexpected
token `)'
../configure: line 7653: `      OSF*)'

This effects both tcl and tk and seems to be bash-3.1 related. The tcl/tk
folks are aware and a fix is in their cvs. In the mean time short of
fiddling with the macros I opted for a sedit of the configure of tcl and tk.

These lines should be removed when tcl/tk are next updated and I have made note
in each of the BUILDs.

Up til now I was unable to compile either on a new install using 1.6.0-rc1 with
bash-3.1. So this temp fix works for me here. 

This may need some further polishing.



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@17987 fe636330-a7f5-0310-8408-fdd8893149bb