OpenHBCI is now outdated and no longer maintained. * configure.in: Release version 0.9.18. Final release. Project is closed.
* src/plugins/keyfile/mediumkeyfile.cpp: Sucessfully finished the
c2-Medium backport -- for c2-Media of version >= 1.6, the
OHBCI_TAG_HEADER is the very first tag instead of the
TAG_VERSION_MAJOR. This medium will now read any c1-Medium and any
c2-Medium, ignoring any tag from format-versions > 1.4. The medium
will write as following: If it reads a c1-Medium, it will write a
c1-Medium with mediumkeyfile-version=1.2. If it reads a c2-Medium,
it will write a c2-Medium with mediumkeyfile-version=1.2 (the new
exponent from version 1.3 is not implemented).



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@18017 fe636330-a7f5-0310-8408-fdd8893149bb