vi no longer dies with a segmentation fault if a line does not fit on the screen after an insertion.
Traditional regular expressions sometimes failed to operate correctly
since the last release.

The argument to the -w option is correctly recognized now.

A non-null exit status is now returned if a file could not be opened, if
an invalid address is given for a command, or if a tag is not found.


and others.



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@18032 fe636330-a7f5-0310-8408-fdd8893149bb