buzz buzz buzz, hype hype hype
For the flashaholics, a little sumptin with freedom attached.

Was not able to get it to build with the docbook stuff. Dunno 
why, tried several things but no luck. Perhaps someone with more
knowledge in that area could help.

The plugin at this time gets thrown in ;

/usr/lib/firefox-1.5.0.1/plugins

For us KDE folks, if you already have firefox installed a simple
rescan in the "Configure the browser plugins" is all it takes. For
other browsers, I've no idea.

It just now dawned on me, a DEPENDS should have at least firefox. I'll
do that here in a minute.



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@18682 fe636330-a7f5-0310-8408-fdd8893149bb