Adding this nss module here. Its in zbeta for a reason.
I tested it with gaim, and it works. However, I have no means
or will to test evolution or any other part that relies on those
headers. Go ahead and do it.

If it works for any other application that currently depends on
mozilla for the nss headers then we might as well use this module
to get rid of those dependencies.



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@19744 fe636330-a7f5-0310-8408-fdd8893149bb