Inital Import: timezone-data
This module provides the /usr/share/zoneinfo/* files and the utilities
that exist for dealing with those timezone files.  Currently our glibc
module is responsible for installing/maintaining those files.  Unfortunatly
glibc's is out of date, especially on the eve of the US's daylight savings
time change, plus other changes thoughout the world.  Having a seperate,
updateable module will be useful when new timezone changes are released.
This follows the lead of other distributions in creating a stand-alone,
updateable timezone package.

The concern at the present time is that these files clobber what glibc
currently installs.  However, do to the restricted view of the MD5SUM 
check in lunar [no]fix, these clobbered files do NOT flag glibc as being
"broken."  So partial good news.  Moe is however working on a new glibc
module, and we will best try to figure out to make these two packages
properly co-exist.

Once this gets all hammered out, it might be good to have these updated
timezone files on the next installer ISO so that after the March DST
time changes take effect, the time will be correct on the ISOs.


git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@23126 fe636330-a7f5-0310-8408-fdd8893149bb