Initial import. I believe all the dependencies are covered. If not let me know.
There is one minor nit. When run /tmp/tmpsong.mid gets created 
with the current user permissions. So if another user logs in to
run jfduke3d, it will not. So you can either delete it or change
it to something like games:games.

Also just so everyone is aware. The duke3d.grp files are not something
that can be freely passed around. All else is GPLed or very end
user flexible.



git-svn-id: svn+ssh://tchan@foo-projects.org/var/svn/lunar/moonbase/trunk@23275 fe636330-a7f5-0310-8408-fdd8893149bb