Version bump to git-1.5.1.6, gtkhtml3-3.14.1, evolution-2.10.1,
evolution-data-server-1.10.1, evolution-exchange-2.10.1. Added gtkhtml312
module as a temporary solution to the gnucash/evolution conflict of requiring
different versions of gtkhtml3. Once gnucash-2.2.0 is released then evolution
and gnucash can go back to using the same versions of gtkhtml3. NOTE: the api
of gtkhtml3 has changed in 3.14.x and is NOT backwards compatible enough for
gnucash. This means the highest version of gtkhtml3 that gnucash can currently
use is 3.12.3. I did check with some other distros and they have gone this
route as well -- namely having two versions of gtkhtml3.


git-svn-id: svn+ssh://foo-projects.org/var/svn/lunar/moonbase/trunk@24487 fe636330-a7f5-0310-8408-fdd8893149bb