While testing a new pcb-cvs module its configure barfed with this error;
checking if /usr/bin/m4 has the division involving negative
numbers bug... yes
configure: error: It appears that /usr/bin/m4 has a bug
involving division with negative numbers.  In particular it
just returned the result that -2/2 = 2147483647 instead of -1.
This is a known bug in GNU m4-1.4.9.  Please install a non-broken m4.

This version of m4 is ok.



git-svn-id: svn+ssh://foo-projects.org/var/svn/lunar/moonbase/trunk@24605 fe636330-a7f5-0310-8408-fdd8893149bb