Initial import for mDNSResponder. This is the Apple one that was adopted by KDE in loo of howl or avahi (sigh). Anyway, kdelibs3 will make use of it. You might have to recompile kdenetwork3, I didn't have to but just in case you have difficulty getting it to work.
To kick it all off. Right click on the kicker panel, click on
"add applet to panel", look for Public File Server and click on 
it. Then on the new icon you have in kicker, right click and start
configuring, you'll need a ~/public_html folder. Umm don't forget
to start the service.


git-svn-id: svn+ssh://foo-projects.org/var/svn/lunar/moonbase/trunk@24621 fe636330-a7f5-0310-8408-fdd8893149bb