Correct some logical mistakes:
Ghostscript has to have cups installed in order to build required printer
backend for it to actually interface with cups. Thus I removed the
dependency of cups on it. People might need to get a little smarter .. but
hey, gutenprint is going to pull in ghostscript anyway.

Note: espgs is deprecated and should be removed.



git-svn-id: svn+ssh://foo-projects.org/var/svn/lunar/moonbase/trunk@25864 fe636330-a7f5-0310-8408-fdd8893149bb