Version bump. Leaving the mozilla-cplusplus.patch in but disabled for the moment. The hunks fail.
For the time being we need to use --without-system-curl. The
current moonbase version causes the build to fail on some ftp
stuff.

Adding binfilter.

At one point the build failed on for file in *.desktop loop.
It bailed on the sedit for the exaile.desktop and the install
in that loop couldn't make the .desktops because they were
already there. I think the change to "$file;" fixed that because
I was not able to duplicate the failure.



git-svn-id: svn+ssh://foo-projects.org/var/svn/lunar/moonbase/trunk@25958 fe636330-a7f5-0310-8408-fdd8893149bb