free cross-platform LaTeX editor [new module]
BUILD file is based on the internal BUILD.sh prompting command supplied.
The DEPENDS probably contains too much: texmaker only needs qt4 to build
but the rest are specified as these are the default commands that are
called at runtime (and are configurable via a Preferences menu)
Also had problems with optional_depends on xpdf because I already had
poppler installed, and the two conflict, and there's no %ALIAS for this.

Other than that, compiles and runs on my box without problem.



From:  duncan.gibson@xs4all.nl


git-svn-id: svn+ssh://foo-projects.org/var/svn/lunar/moonbase/trunk@26684 fe636330-a7f5-0310-8408-fdd8893149bb