mono, mono-basic, libgdiplus: bump the version to 1.2.6.
Some of the highlights for this release include:

-Native Windows.Forms driver for MacOS X allows Winforms-based
applications to run without an X server.
-Support for the ASP.NET AJAX APIs and controls.
-Support for FastCGI deployments: ASP.NET can now be deployed on a
multitude of servers that implement the FastCGI protocol (lighttpd
for example) in addition to Apache.
-Windows.Forms now supports the WebControl on Windows and Linux using
Mozilla.
-Runtime will now consume much less memory for 2.0-based applications
due to various optimizations in generics support as well as including
many new performance improvements and an updated verifier and an
implementation of CoreCLR security.
-C# compiler is quickly approaching full 3.0 support, most of the
basics work right now (except support for System.Query.Expression AST
generation).
-Mono 1.2.6 can now be used as an SDK for creating Silverlight 1.1
applications on all platforms. This allows developers to create
applications that target Silverlight without requiring a Windows
installation.