amarok: Ruby is now a hard dependency.
The lin appears to go fine but amarok never shows up in kmenu. If you look
through the compile log you'll see a message amarok will not be built if
ruby is not present. This is something I noticed several months back but
forgot about it til today when doing a fresh install.