sudo: version bump to 1.6.9p11
This release fixes a compilation problem on SCO UNIX related to how
it stores the high resolution timestamps in struct stat. The URI
specifier can now be used in ldap.conf even when the LDAP SDK doesn't
support ldap_initialize(). There is a new %p prompt escape that
expands to the user whose password is being prompted, as specified by
the rootpw, targetpw, and runaspw sudoers flags.