scribus: move it to the qt-apps dir and version bumped to 1.3.3.10.
Several fixes and improvements to text frames and the Story Editor.
A new Arabic Translation. More translation and documentation updates.
Many improvements to PDF Forms exporting and non-Latin script
handling in PDFs. Several fixes to protect against possible crashes.
Improved Color Managed Display in some cases. Some fixes to the
Scripting plugin. One of the major additions to this release is the
final complete German translation of the Scribus documentation.