e2fsprogs: version bump to 1.40.5.
Changes:
Fix a potential overflow big in e2image if the device name
is too long.

Mke2fs will now create new filesystems with 256 byte inodes
and the ext_attr feature flag by default.

Teach e2fsck to ignore certain "safe" filesystem features
which are set automatically by the kernel.

Add support in tune2fs and mke2fs for making a filesystem
as being "ok to be used with test kernel code".

Change e2fsck -fD so that it sorts non-htree directories by
inode numbers instead of by name, since that optimizes
performances much more significantly.

If e2image fills the disk, fix it so it exits right away
instead of spewing large numbers of error messages.

If ftruncate64() is not available for resize2fs, let it use
ftrucate() instead, but have it check to see if the size
would get truncated, and skip calling ftruncate in that
case.

Add support for detecting HFS+ filesystems in the blkid
library.

Add support in the blkid library for ext4/ext4dev
filesystems.

Fix a bug in blkid where it could die on a floating point
exception when presented with a corrupt reiserfs image.

Fix blkid's handling of ntfs UUID's so that leading zeros
are printed such that UUID string is a fixed length.

Add sample python bindings for the uuid library to the
contrib directory.

Fix debugfs's 'lsdel' command so it uses ext2fs_block_iterate2
so it will work with large files.

Allow the debugfs 'undel' command to undelete an inode
without linking it to a specific destination directory,
since this might require allocating blocks that could
overwrite some yet-to-be-recovered deleted files.