ntfs-3g: version bump to 1.2129.
The release has FUSE-LITE built-in, so only the FUSE kernel module is
required. The driver now works with the uClibc library. A new utility
called ntfs-3g.probe was added to probe a volume for mountability.
Unique exit codes are used for ntfs-3g and ntfs-3g.probe to indicate
the volume state. A fix was made for the case when a SIGTERM signal
may cause a deadlock that could block the shutdown process.