vsftpd: version bump to 2.0.6.
STOU support was fixed. Bugs in the xferlog format were fixed.
OPTS UTF-8 ON is now recognized, but nothing is done with it.
Firefox's attempts to RETR directories no longer appear as
download attempts in the log. There were other minor fixes
and tweaks. SSL support was enhanced; it is now possible to
insist that FTP clients send client certs, which in turn
enhances the data connection security.