hdparm: version bump to 8.1.
A major code overhaul was made. Support for libata SATA/PATA
drives was significantly expanded. Much better compliance
with the SCSI-ATA-Translation specification was achieved.
The -A, -W, and -M flags can now be used without a parameter
to query the current setting. Several enhancements to the
-I information were made. New flags "--make-bad-sector",
"--repair-sector", and "--read-sector" were added. The "-N"
flag was added to get/set max address/HPA. Many other fixes
and enhancements were made.