sudo: version bump to 1.6.9p13.
Sudo will now set the nproc resource limit to unlimited on
Linux systems to work around Linux's setuid() resource limit
semantics. On PAM systems the resource limits will be reset
by pam_limits.so before the command is executed. Sudo now
has SELinux support that can be used to implement role based
access control (RBAC). A role and (optional) type may be
specified in sudoers or on the command line. These are then
used in the security context that the command is run as.
This release also fixes a Kerberos 5 compilation problem
with MIT Kerberos.