smb4k: version bump to 0.9.3.
This version comes with a backport from the current
development branch that improves the handling of login
information. To achieve this, the format of the data that
is written to the wallet had to be changed and, thus,
Smb4K 0.9.3 is not compatible with earlier versions. The
conversion of the old data will be done automatically.
Apart from that, two (un)mount issues and a regression in
the network browser have been eliminated.