syslinux : version bump to 3.62.
The primary new feature of this release is support for
hierarchical submenus in the simple menu system. In addition
to that, this release eliminates a lot of hardcoded
limitations, such as the maximum number of labels, which
are now only limited by memory, and the length of the
command line. Furthermore, this release fixes the graphical
display on systems which do not support the linear framebuffer
option in the VESA BIOS.