dbus: Solving the XML renderer madness once and for all.
First of all, expat is a pretty basic package. Its tiny compared to libxml2 and takes little to none time to compile. Plus, its also being used by hal .. which has become as vital as dbus to any linux system these days.

And secondly, this warning confined in the configure script of dbus is pretty obvious:

WARNING: You have chosen to use libxml as your xml parser however this code path is not maintained by the D-Bus developers and if it breaks you get to keep the pieces.  If you have selected this option in err please reconfigure with expat (e.g. --with-xml=expat).

Therefore, from now on, we'll be using expat excusively. At least with this module.