scummvm: version bump to 0.11.1.
Changes:
Improvements for the Digital iMUSE subsystem. This fixes
several glitches in The Curse of Monkey Island.
Fixes for cursors in games by Humongous Entertainment.
Fix for zombies in King's Quest 4.
Fix for changing palettes in fanmade games using AGIPAL.
Fixed some conversation crashes in the German version of
Lure of the Temptress.
Fixed operation of the optional copy protection dialog in
the German version of Lure of the Temptress.
Added saving of conversation flags in Lure of the Temptress
as to whether a particular conversation option had been
previously selected or not.
Fix for the rat maze bug in Inherit the Earth which made
game not completable.
Fixes for Inherit the Earth and I Have no Mouth, and I Must
Scream game startup on a number of platforms.
More bug fixes.