tellico: version bump to 1.3.1.
Data sources for discogs.com and Google Scholar were added.
The CrosssRef source was updated to grab more metadata.
DOI search was added to z39.50 and SRU sources.
Loading performance was improved to delay loading linked
images.
The BoardGameGeek source was updated to grab the cover image.
A bug that prevented BibTeX format from working for
external scripts was fixed.
The HTML export was fixed to correctly link to entries with
multiple titles.