kdebindings4: changing the optional_depends to PyQt-4
other wise the PyKDE4 python bindings do not get built.
During the build it will set at 100% for what seems an
eternity causing you think something has gone awry. Have
patience.

Also, while there is a PyKDE4 package at riverbank we IMO
do not need to add it as a separate module. According to
riverbank they and the kde folks keep the two in sync or
nearly so.