pciutils: version bump to 3.0.0.
The lspci utility is now able to use a central database to
look up names of devices that are missing from the local pci.ids
file. This is accomplished by DNS queries on a special zone.
Also, the libpci library can be now built as a shared library
with a stable ABI.