openoffice-src: version bump.
This works on x86_64 if you choose j2sdk, see its DETAILS.x86_64.

There is some jiggling/simplification, etc that could be done here
where x86 and x86_64 could be merged into one set of scripts but I
iz tired atm. So if there is anyone that wants to diddle with the
x86 side of things.... feel free and enjoy.