transcode: Some fleshing out of DEPENDS,
The configure notes: "internal libac3 decoder is deprecated and
will be removed in future releases" so went ahead and changed
the default.

Added --enable-libpostproc to the BUILD and since ffmpeg defaults
to postprocessing adding it should cause no issues.

The lzo addition needed a fuller path specified since the configure
isn't able to walk down the tree.

As for libfame, well I have no resolution atm.