connman: a new network connection manager
This includes required dependencies connman-gnome, libgdbus, resolvconf.

connman is a background daemon that runs as root. the userspace connman-applet
runs in the systray and gives a configuration interface to the user.

this interface can be used to manage WPA and the daemon will choose to
enable the best link available where ever possible.

connman requires dhclient so 'dhcp' package is needed. resolvconf is a
simple debian package to mitigate /etc/resolv.conf changes. The whole
set of programs heavily uses dbus allthough they're planning on making
a dbus-less version as well for embedded. obviously wpa_supplicant
is required for managing wireless connections.

all in all this is a great addition for the xfce desktop and can replace airconfig
where used, so Brian can stop messing with it and do real work ;)