smb4k: version bumped to 0.9.5.
This release fixes a logon issue that was introduced in
version 0.9.4: smb4K sent the login and password to the
master browser by default when asking for the browse list.
This could cause at least master browsers running Samba to
return an empty browse list. This behavior has now been made
configurable and can be enabled in the configuration dialog
under "Network | Authentication | Master browsers need
authentication". It is turned off by default.