nasm: version bumped to 2.03.
This release adds support for Intel AVX, CLMUL, and FMA
instructions, including YMM registers as well as the INVEPT,
INVVPID, and MOVBE instructions. It also adds support for
the DWARF debugging format for ELF targets, adds SEH support
for Windows, improves generation of Makefile dependencies,
and fixes a large number of bugs.