aqbanking: version bumped to 3.5.1.
This release fixes some more bugs and can be built with the
special version of aqbanking-cli with EBICS support. When a
new PIN is to be entered for any reason (e.g. changing the
password of a key file), the user is now always asked (no
PIN caching in this case).