fox: version bumped to 1.6.34.
Changes:
Fixed bug in FXSettings; symptoms: every once in a while, a key-value pair was skipped.
Small update to MAC-OSX implementation of FXThread.
Fixed some minor things in fxutils for particular Windows header files.
On Windows, block widget size update if it is minimized; minimized is a state, and does not cause a new size.
Minor tweak to order in which size constraints are being passed to Window Managers on X11. Some Window Managers had a problem with the old order.