konversation: version bumped to 1.1.
The additions and improvements in this release are both user-visible and under
the hood. Some of the highlights are rewritten connection handling (robustness
and correctness improvements, better support for IRC URLs, bookmarking and
more), redone DCC with better UI and Passive/Reverse DCC support, a redone away
system with the addition of auto-away support, redone and much more useful
remember / marker line support, a new outbound traffic scheduler that is capable
of aggressive throttling to avoid flooding while smartly reordering messages to
improve latencies, great convenience additions like a "Next Active Tab" shortcut,
and much, much more, along with a large number of bugfixes and tweaks to round
things out.

This is the last KDE 3.x version.