gambas2: version bumped to 2.8.0.
This release compiles correctly with gcc 4.3.
When displaying a method signature, the current argument
is underlined. IDE shortcuts of different menus should not
interfere anymore.
The GTK+ component can deal with font underline and
strikeout in many places. The Qt component tray icons are
now visible inside Gnome or XFCE.