firefox: fixing an exceedingly unusual bug
But it just happened to catch me today.  Somehow I had more than one old
version of Firefox installed, and this caused the error message:

/var/lib/lunar/moonbase/firefox/POST_INSTALL: line 5: [: too many arguments

when I went to lin the latest version of Firefox.

This is an easy enough problem to solve, so I fixed it so if there are
multiple old versions, it'll clean up all of their /usr/lib/firefox-$ver
directories.