ntp: version bumped to 4.2.4p5.
This release fixes a number of Windows-specific ntpd bugs and
platform-independent ntpdate bugs. A logging bugfix has been applied to the
ONCORE driver. The "dynamic" keyword is now obsolete and deferred binding to
local interfaces is the new default. The minimum time restriction for the
interface update interval has been dropped. A number of minor build system
and documentation fixes are included.