gambas2: version bumped to 2.8.2.
Changes:
The bits manipulation functions now work correctly.
The tray icons are correctly refreshed when the icon has changed.
The interpreter is now linked with the thread library on Linux systems.
When opening a project in the IDE, the packager information is correctly reset.
The Simplified Chinese translation has been completed.