BeutifulSoup: version bumped to 3.0.7a.
Changes:
Fixed a UnicodeDecodeError when unpickling documents that contain non-ASCII characters.
Fixed a TypeError that occured in some circumstances when a tag contained no text.
Jump through hoops to avoid the use of chardet, which can be extremely slow in some circumstances. UTF-8 documents should never trigger the use of chardet.
Whitespace is preserved inside <pre> and <textarea> tags that contain nothing but whitespace.
Beautiful Soup can now parse a doctype that's scoped to an XML namespace.