nasm: version bumped to 2.04.
This is primarily a bugfix release. In particular, it should end once
and for all abnormalities introduced by forward references when
pass-limited. It does introduce some new features, mostly in the
preprocessor. It adds support for packed BCD and UTF-16/UTF-32
constants. It also adds support for "smart" alignment, padding the
bypassed region with optimized code.