cairo: version bumped to 1.8.0.
Many improvements were made to Cairo's ability to handle text, including
a new "user fonts" feature as well as allowing glyphs to be embedded in
PDF output along with their original text (for searching, selection, and
copy-and-paste). A major revamp of the build system makes it much easier
to build Cairo on various platforms. The performance on remote X servers
without the Render extension was improved. The output of cairo-ps and
cairo-pdf was made more efficient. Completions were made to the API. Many
bugs were fixed.