NVIDIA: version bumped to 177.70.14.
Changes:
Added support for GeForce 8600 GS.
Fixed a problem with missing rendering in OpenGL Workstation Overlays.
Fixed a problem with running some SDL applications and virtual terminal switching.
Fixed a potential crash in nvidia-settings when saving to the X configuration file.
Improved error recovery paths in the case of corruption of the commands sent to the GPU.