poppler: version bumped to 0.10.0.
Chamges:
Initial JavaScript support was added.
Improvements were made to Annotation and in the Arthur and Cairo based renderers.
A JPEG2000 decoder based on OpenJPEG was added.
JPEG rendering when not using the libjpeg based decoder was fixed.
An out of memory condition on documents that specify huge fonts was fixed.
Adobe Reader's behavior is emulated on documents with duplicate keys in Dictionaries.
Forms improvements and movie fixes were made.
Many improvements were made to the Qt and Glib frontends, the utilities, and the documentation.
Several other improvements and bugfixes were made.