libev: version bumped to 3.48.
This release mainly adds additional checks to the epoll backend to work
around the design bugs in the epoll kernel interface, by detecting and
ignoring events sent to the wrong process or otherwise not being
current. It also tries even harder to reduce the number of syscalls for
common use patterns in the epoll backend. The documentation has once
again been expanded and improved, and a number of minor issues have been
fixed, among them that ev_default_destroy would leave libev in a state
where one couldn't create the default loop again.