sudo: version bumped to 1.6.9p18.
Changes:
A crash which occurred when the -i flag was used with a uid not in the password database has been fixed.
Sudo now operates in the C locale again when doing a match against sudoers.
A potential crash when a glob matches a large number of files has been fixed.
A compilation problem with certain versions of Heimdal KerberosV has been fixed.
When setting the umask, sudo will now use the union of the user's umask and the value set in sudoers.
Sudo now stats fewer files when doing a wildcard match.