libev: version bumped to 3.49.
The only major change is that inotify is now used even on older
kernels that are known to be buggy, but only as a hint. Also,
libev now has a small list of "known to be good" filesystems where
inotify is known to work, and will use inotify as a hint only on
other filesystems (such as NFS). The documentation has been updated
regarding the dangers of using stat watchers on networked file
systems.