dmidecode: version bumped to 2.10.
SMBIOS 2.6 support was added. This includes new enumerated values for
recent hardware, as well as bus address for system slots and a new
entry type for on-board devices. The ability to dump the SMBIOS table
to a small binary file, and to read the SMBIOS table back from such
binary files, was added.