squirrelmail: version bumped to 1.4.17.
White space wrapping of auto-generated SquirrelMail option widgets may now be
controlled. Matching of alternate identities when replying was fixed. HTTPS
detection under Windows IIS was fixed, as it was incorrectly setting cookies
to be transmitted only over a secure connections when none existed. An XSS
exploit in hyperlinks when rendering messages was fixed.