x11-vnc: version bumped to 0.9.6.
x11vnc's SSL encryption is extended to support the VeNCrypt and TLS VNC
security types. An increasing number of viewers support the former.
Anonymous Diffie-Hellman key exchange support is provided, and Certificate
Revocation Lists are implemented. The Java viewer applet is downloadable
via the VNC port as well as the HTTP port. If x11vnc is not built with
the Avahi Zeroconf library, an external helper program is used instead. A
launcher for naive users is provided in x11vnc.desktop.