sqlite: version bumped to 3.6.7.
Changes:
Reorganize the Unix interface in os_unix.c
Added support for "Proxy Locking" on MacOSX.
Changed the prototype of the sqlite3_auto_extension() interface in a
way that is backwards compatible but which might cause warnings in new
builds of applications that use that interface.
Changed the signature of the xDlSym method of the sqlite3_vfs object in
a way that is backwards compatible but which might cause compiler
warnings.
Added superfluous casts and variable initializations in order to suppress
nuisance compiler warnings.
Fixes for various minor bugs.