SDL_gfx: version bumped to 2.0.18.
Changes:
A libtool update for dependencies.
MSVC C89 variable hoisting to enable MSVC compilation.
A fix for uninitialized variables in _putPixelAlpha().
Reversion to the original alpha routine in _putPixelAlpha(), as the new one was causing issues in 16bbp.
A colorkey handling patch for rotozoom.
Optimization has been downgraded to -O for asm compatibility with gcc4.