BeautifulSoup: bumped version and fixed source URL
If you download just the file $SOURCE_URL/BeautifulSoup.tar.gz, that
gets you the very latest version, resulting in an SHA1sum mismatch.

If you download just the file $SOURCE_URL/BeautifulSoup-$VERSION.tar.gz,
that works only if you're downloading the very latest version of
BeautifulSoup.  Previous versions (and not the latest) are in
$SOURCE_URL/3.x, which is why I set $SOURCE_URL[0] and $SOURCE_URL[1].

Why he did it this way is beyond me.  Possibly to annoy maintainers of
Linux distros.