ntsf-3g: version bumped to 2009.1.1.
Built-in, transparent UTF-8 conversion support was added.
This solves problems with hidden and inaccessible filenames having
national characters.
Support for getting or setting the file creation timestamp on Mac
OS X was added.
A problem was fixed in which an extended attribute got corrupted if
the new attribute size was smaller than the original size.
A memory leak when an extended attribute or alternate data stream
couldn't be opened was fixed.
An error is reported if a filename cannot be listed while reading a
directory.