libchipcard: version bumped to 4.2.4.
This release fixes a few bugs and adds descriptions for some more readers.
and reintroduces a CryptToken for STARCOS cards and improves
support for the upcoming "Elektronische Gesundheitskarte" which can now be
read by the tool kvkread along with old "Krankenversichertenkarte" cards.
Hardware scans can now be triggered by signals again (was broken in previous
releases).